requirements.txt
Django==1.4.2
Jinja2==2.6
PIL==1.1.7
PyJWT==0.1.2
Pygments==1.5
South==0.7.6
Sphinx==1.1.3
distribute==0.6.31
dj-database-url==0.2.1
django-classy-tags==0.3.4.1
django-cms==2.3.4
django-geoposition==0.1.4
django-jsonfield==0.8.12
django-mptt==0.5.2
django-picklefield==0.2.1
django-positions==0.4.3
django-reversion==1.6.4
django-sekizai==0.6.1
django-sms-gateway==1.2.0
django-twilio==0.4
django-uuidfield-2==0.6.5
docutils==0.9.1
flickrapi==1.4.2
geopy==0.94.2
html5lib==0.95
httplib2==0.7.7
psycopg2==2.4.5
python-dateutil==2.1
simplejson==2.6.2
six==1.2.0
sorl-thumbnail==3.2.5
twilio==3.4.1
wsgiref==0.1.2
